Fix the openconnect_services script
It turns out that run_rc_script does not support passing more than one
argument to the specified script. The openconnect script received
the command (e.g., "onestop" or "start") but not the name of the service
(e.g., "default" or "myvpn"). A simple fix is to not use run_rc_script
at all and just run the openconnect script directly.

This patch fixes issue #1.

Reported by:	swills