Use daemon(8) to daemonize openconnect
It turns out that openconnect(8) does not daemonize correctly. The issue
seems to be related to the fact that a SIGHUP is sent to the process
when it's orphaned, which is not handled by openconnect(8) as we would
expect.

The solution here is to use daemon(8) which can take care of this and
other potential bugs we could encounter in the future. Unfortunately,
setting command=/usr/sbin/daemon in an rc(8) script has its own problems
like breaking the standard *_flags variable for the openconnect
services.

This patch resolves issue #2.

Reported by:	swills